refactor(navigation): move navigation pages to api package
